+++ NEWS.md
@@ -8,6 +8,8 @@
 
 #### 🍬 miscellaneous
 
+  * Eliminates {dplyr} dependency (#12)
+
 
 # GSODTools 1.0.5 (2022-06-08)